requirements.txt
Pillow==9.3.0
qrcode==7.3.1
